Final touches I guess
Increase the mobile padding of <body>
Increase the font-size and reduce the line-height of .main-text__title
Use box-shadow and not border for focus ring. Why? border is part of the height when box-sizing: border-box is set. Thus using a border pushes other elements down when an input is fouced on. This change cancels the need for the [aria-invalid] selector.